+++ client/src/pages/signal-generator/examples/README.md
@@ -1,3 +1,3 @@
 To add a new example python code of a signal, create a new file in this dir with a name describing the signal and .js extension.
 Use one of the others as an example, making sure the name that comes after `export const` matches the filename.
-Lastly, in exampleList.js, add your new file as both an import and to the `examplesList` list.
+Lastly, in example-list.js, add your new file as both an import and to the `examplesList` list.
