+++ client/src/pages/signal-generator/examples/example-list.ts
