+++ client/src/pages/signal-generator/signal-generator.tsx
@@ -5,7 +5,7 @@ import React, { useEffect, useState, useCallback } from 'react';
 import { vscodeDark } from '@uiw/codemirror-theme-vscode';
 import CodeMirror from '@uiw/react-codemirror';
 import { python } from '@codemirror/lang-python';
-import { examplesList } from '@/pages/signal-generator/examples/exampleList';
+import { examplesList } from '@/pages/signal-generator/examples/example-list';
 
 export default function SignalGenerator(props) {
   const [state, setState] = useState({
